:truck: change the plugins into utils
:warning: this pull request needs new migrations to database because the database has been changed (in goldennum/views.py)
Why to make this pull request? - because in Django, all other applications are "plugins“ and to make a "plugin of plugin" is duplicated, it is better to merge it into the original goldennum floder and become a util function.
What will it do?  - making it a part of the project makes it possible in the future to change the data inside the django admin system, which made it more secure and easy to use.